fs-viewer: don't close things
For some reason I'm getting errors about them already being closed. This
is a bit leaky, but it's a demo so I'm inclined to just say to heck with
it for now. I'd like to know what's going on, but @zombiezen has said
that the finalization story for Clients is pretty shaky anyhow, so my
inclination is to wait for go-capnproto v3 before fussing about it.