Mostly port to sandstorm-http-bridge.
I'm planning on using this partially to validate that the work I've been
doing on the bridge actually covers the needed functionality. It seems
to partially work, but there's some debugging of the bridge that needs
doing still.

The replace directive points at the version of go.sandstorm I have
locally. That will have to get removed at some point, but for now it's
convenient.